xattrs for MPS: do not fail if removing security.selinux label failed
Resolves Redmine #4115 for MPS MIA.

During partition synchronization, if the source partition has an
unlabelled file (no "security.selinux" extended attribute), rsync would
try to remove a default label from target partition and fail. If the
SELinux is enabled, the attribute cannot be removed by design and
nothing can be done about that.

Therefore in order not to break the TSP install, since this is expected
situation, ignore this error.